+++ libnavui-dropin/src/main/java/com/mapbox/navigation/dropin/map/MapViewOwner.kt
@@ -2,13 +2,21 @@ package com.mapbox.navigation.dropin.map
 
 import com.mapbox.maps.MapView
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import java.util.concurrent.CopyOnWriteArraySet
 
 @OptIn(ExperimentalPreviewMapboxNavigationAPI::class)
 internal class MapViewOwner {
 
     private var mapView: MapView? = null
+        set(value) {
+            field = value
+            _mapViews.value = value
+        }
     private val listeners = CopyOnWriteArraySet<MapViewObserver>()
+    private val _mapViews = MutableStateFlow(mapView)
+    val mapViews = _mapViews.asStateFlow()
 
     fun registerObserver(observer: MapViewObserver) {
         if (listeners.add(observer)) {