+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/customnavview/MapboxNavigationViewCustomizedActivity.kt
@@ -49,6 +49,7 @@ import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationObserver
 import com.mapbox.navigation.core.trip.session.LocationMatcherResult
 import com.mapbox.navigation.core.trip.session.LocationObserver
+import com.mapbox.navigation.dropin.MapboxEnablableButtonParams
 import com.mapbox.navigation.dropin.MapboxExtendableButtonParams
 import com.mapbox.navigation.dropin.ViewOptionsCustomization.Companion.defaultRouteArrowOptions
 import com.mapbox.navigation.dropin.ViewOptionsCustomization.Companion.defaultRouteLineOptions
@@ -336,6 +337,12 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
             menuBinding.spinnerProfile,
             viewModel.routingProfile,
             ::setRoutingProfile,
+        )
+
+        bindSwitch(
+            menuBinding.toggleEnableCompass,
+            viewModel.enableCompass,
+            ::toggleEnableCompass
         )
     }
 
@@ -687,6 +694,17 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
         }
         binding.navigationView.setRouteOptionsInterceptor { defaultBuilder ->
             defaultBuilder.layers(null).applyDefaultNavigationOptions(routingProfile)
+        }
+    }
+
+    private fun toggleEnableCompass(enabled: Boolean) {
+        binding.navigationView.customizeViewStyles {
+            compassButtonParams = MapboxEnablableButtonParams(
+                enabled,
+                ViewStyleCustomization.defaultCompassExtendableButtonParams(
+                    this@MapboxNavigationViewCustomizedActivity
+                )
+            )
         }
     }